+++ src/billing/Project.js
@@ -1,75 +1,29 @@
-import { Icon, Modal, SpinnerOverlay } from '@terra-ui-packages/components';
+import { SpinnerOverlay } from '@terra-ui-packages/components';
 import _ from 'lodash/fp';
 import * as qs from 'qs';
 import { Fragment, useEffect, useMemo, useState } from 'react';
 import { div, h, p, span } from 'react-hyperscript-helpers';
-import * as Auth from 'src/auth/auth';
+import { BillingAccountControls } from 'src/billing/BillingAccount/BillingAccountControls';
+import { BillingAccountSummary } from 'src/billing/BillingAccount/BillingAccountSummary';
 import { Members } from 'src/billing/Members/Members';
 import { ExternalLink } from 'src/billing/NewBillingProjectWizard/StepWizard/ExternalLink';
 import { SpendReport } from 'src/billing/SpendReport/SpendReport';
-import { billingRoles } from 'src/billing/utils';
-import { getBillingAccountIconProps } from 'src/billing/utils';
+import { accountLinkStyle, billingRoles } from 'src/billing/utils';
 import { Workspaces } from 'src/billing/Workspaces/Workspaces';
-import { ButtonPrimary, IdContainer, Link, VirtualizedSelect } from 'src/components/common';
+import { Link } from 'src/components/common';
 import { DeleteUserModal, EditUserModal, NewUserModal } from 'src/components/group-common';
 import { icon } from 'src/components/icons';
 import { InfoBox } from 'src/components/InfoBox';
-import { TextInput } from 'src/components/input';
-import { MenuButton } from 'src/components/MenuButton';
-import { MenuTrigger } from 'src/components/PopupTrigger';
 import { SimpleTabBar } from 'src/components/tabBars';
 import { Ajax } from 'src/libs/ajax';
 import { isAzureBillingProject, isGoogleBillingProject } from 'src/libs/ajax/Billing';
 import colors from 'src/libs/colors';
 import { reportErrorAndRethrow } from 'src/libs/error';
 import Events, { extractBillingDetails } from 'src/libs/events';
-import { FormLabel } from 'src/libs/forms';
 import * as Nav from 'src/libs/nav';
 import { useCancellation, useGetter, useOnMount, usePollingEffect } from 'src/libs/react-utils';
-import { contactUsActive } from 'src/libs/state';
 import * as StateHistory from 'src/libs/state-history';
-import { topBarHeight } from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
-
-const accountLinkStyle = { color: colors.dark(), fontSize: 14, display: 'flex', alignItems: 'center', marginTop: '0.5rem', marginLeft: '1rem' };
-
-const BillingAccountSummaryPanel = ({ counts: { done, error, updating } }) => {
-  const StatusAndCount = ({ status, count }) =>
-    div({ style: { display: 'float' } }, [
-      div({ style: { float: 'left' } }, [Icon(getBillingAccountIconProps(status))]),
-      div({ style: { float: 'left', marginLeft: '0.5rem' } }, [`${status} (${count})`]),
-    ]);
-
-  const maybeAddStatus = (status, count) => count > 0 && div({ style: { marginRight: '2rem' } }, [h(StatusAndCount, { status, count })]);
-
-  return div(
-    {
-      style: {
-        padding: '0.5rem 2rem 1rem',
-        position: 'absolute',
-        top: topBarHeight,
-        right: '3rem',
-        width: '30rem',
-        backgroundColor: colors.light(0.5),
-        boxShadow: '0 2px 5px 0 rgba(0,0,0,0.25)',
-      },
-    },
-    [
-      div({ style: { padding: '1rem 0' } }, 'Your billing account is updating...'),
-      div({ style: { display: 'flex', justifyContent: 'flex-start' } }, [
-        maybeAddStatus('updating', updating),
-        maybeAddStatus('done', done),
-        maybeAddStatus('error', error),
-      ]),
-      error > 0 &&
-        div({ style: { padding: '1rem 0 0' } }, [
-          'Try again or ',
-          h(Link, { onClick: () => contactUsActive.set(true) }, ['contact us regarding unresolved errors']),
-          '.',
-        ]),
-    ]
-  );
-};
 
 const groupByBillingAccountStatus = (billingProject, workspaces) => {
   const group = (workspace) =>
@@ -89,259 +43,6 @@ const groupByBillingAccountStatus = (billingProject, workspaces) => {
 };
 
 const spendReportKey = 'spend report';
-
-const GcpBillingAccountControls = ({
-  authorizeAndLoadAccounts,
-  billingAccounts,
-  billingProject,
-  isOwner,
-  getShowBillingModal,
-  setShowBillingModal,
-  reloadBillingProject,
-  setUpdating,
-}) => {
-  const [showBillingRemovalModal, setShowBillingRemovalModal] = useState(false);
-  const [showSpendReportConfigurationModal, setShowSpendReportConfigurationModal] = useState(false);
-  const [selectedBilling, setSelectedBilling] = useState();
-  const [selectedDatasetProjectName, setSelectedDatasetProjectName] = useState(null);
-  const [selectedDatasetName, setSelectedDatasetName] = useState(null);
-
-  const signal = useCancellation();
-
-  // Helpers
-  const setBillingAccount = _.flow(
-    reportErrorAndRethrow('Error updating billing account'),
-    Utils.withBusyState(setUpdating)
-  )((newAccountName) => {
-    Ajax().Metrics.captureEvent(Events.billingChangeAccount, {
-      oldName: billingProject.billingAccount,
-      newName: newAccountName,
-      ...extractBillingDetails(billingProject),
-    });
-    return Ajax(signal).Billing.changeBillingAccount({
-      billingProjectName: billingProject.projectName,
-      newBillingAccountName: newAccountName,
-    });
-  });
-
-  const removeBillingAccount = _.flow(
-    reportErrorAndRethrow('Error removing billing account'),
-    Utils.withBusyState(setUpdating)
-  )(() => {
-    Ajax().Metrics.captureEvent(Events.billingRemoveAccount, extractBillingDetails(billingProject));
-    return Ajax(signal).Billing.removeBillingAccount({
-      billingProjectName: billingProject.projectName,
-    });
-  });
-
-  const updateSpendConfiguration = _.flow(
-    reportErrorAndRethrow('Error updating spend report configuration'),
-    Utils.withBusyState(setUpdating)
-  )(() =>
-    Ajax(signal).Billing.updateSpendConfiguration({
-      billingProjectName: billingProject.projectName,
-      datasetGoogleProject: selectedDatasetProjectName,
-      datasetName: selectedDatasetName,
-    })
-  );
-
-  // (CA-1586) For some reason the api sometimes returns string null, and sometimes returns no field, and sometimes returns null. This is just to be complete.
-  const billingProjectHasBillingAccount = !(billingProject.billingAccount === 'null' || _.isNil(billingProject.billingAccount));
-  const billingAccount = billingProjectHasBillingAccount ? _.find({ accountName: billingProject.billingAccount }, billingAccounts) : undefined;
-
-  const billingAccountDisplayText = Utils.cond(
-    [!billingProjectHasBillingAccount, () => 'No linked billing account'],
-    [!billingAccount, () => 'No access to linked billing account'],
-    () => billingAccount.displayName || billingAccount.accountName
-  );
-
-  return h(Fragment, [
-    Auth.hasBillingScope() &&
-      div({ style: accountLinkStyle }, [
-        span({ style: { flexShrink: 0, fontWeight: 600, fontSize: 14, margin: '0 0.75rem 0 0' } }, 'Billing Account:'),
-        span({ style: { flexShrink: 0, marginRight: '0.5rem' } }, billingAccountDisplayText),
-        isOwner &&
-          h(
-            MenuTrigger,
-            {
-              closeOnClick: true,
-              side: 'bottom',
-              style: { marginLeft: '0.5rem' },
-              content: h(Fragment, [
-                h(
-                  MenuButton,
-                  {
-                    onClick: async () => {
-                      if (Auth.hasBillingScope()) {
-                        setShowBillingModal(true);
-                      } else {
-                        await authorizeAndLoadAccounts();
-                        setShowBillingModal(Auth.hasBillingScope());
-                      }
-                    },
-                  },
-                  ['Change Billing Account']
-                ),
-                h(
-                  MenuButton,
-                  {
-                    onClick: async () => {
-                      if (Auth.hasBillingScope()) {
-                        setShowBillingRemovalModal(true);
-                      } else {
-                        await authorizeAndLoadAccounts();
-                        setShowBillingRemovalModal(Auth.hasBillingScope());
-                      }
-                    },
-                    disabled: !billingProjectHasBillingAccount,
-                  },
-                  ['Remove Billing Account']
-                ),
-              ]),
-            },
-            [
-              h(Link, { 'aria-label': 'Billing account menu', style: { display: 'flex', alignItems: 'center' } }, [
-                icon('cardMenuIcon', { size: 16, 'aria-haspopup': 'menu' }),
-              ]),
-            ]
-          ),
-        getShowBillingModal() &&
-          h(
-            Modal,
-            {
-              title: 'Change Billing Account',
-              onDismiss: () => setShowBillingModal(false),
-              okButton: h(
-                ButtonPrimary,
-                {
-                  disabled: !selectedBilling || billingProject.billingAccount === selectedBilling,
-                  onClick: () => {
-                    setShowBillingModal(false);
-                    setBillingAccount(selectedBilling).then(reloadBillingProject);
-                  },
-                },
-                ['Ok']
-              ),
-            },
-            [
-              h(IdContainer, [
-                (id) =>
-                  h(Fragment, [
-                    h(FormLabel, { htmlFor: id, required: true }, ['Select billing account']),
-                    h(VirtualizedSelect, {
-                      id,
-                      value: selectedBilling || billingProject.billingAccount,
-                      isClearable: false,
-                      options: _.map(({ displayName, accountName }) => ({ label: displayName, value: accountName }), billingAccounts),
-                      onChange: ({ value: newAccountName }) => setSelectedBilling(newAccountName),
-                    }),
-                    div({ style: { marginTop: '1rem' } }, [
-                      'Note: Changing the billing account for this billing project will clear the spend report configuration.',
-                    ]),
-                  ]),
-              ]),
-            ]
-          ),
-        showBillingRemovalModal &&
-          h(
-            Modal,
-            {
-              title: 'Remove Billing Account',
-              onDismiss: () => setShowBillingRemovalModal(false),
-              okButton: h(
-                ButtonPrimary,
-                {
-                  onClick: () => {
-                    setShowBillingRemovalModal(false);
-                    removeBillingAccount(selectedBilling).then(reloadBillingProject);
-                  },
-                },
-                ['Ok']
-              ),
-            },
-            [div({ style: { marginTop: '1rem' } }, ["Are you sure you want to remove this billing project's billing account?"])]
-          ),
-      ]),
-    Auth.hasBillingScope() &&
-      isOwner &&
-      div({ style: accountLinkStyle }, [
-        span({ style: { flexShrink: 0, fontWeight: 600, fontSize: 14, marginRight: '0.75rem' } }, 'Spend Report Configuration:'),
-        span({ style: { flexShrink: 0 } }, 'Edit'),
-        h(
-          Link,
-          {
-            tooltip: 'Configure Spend Reporting',
-            style: { marginLeft: '0.5rem' },
-            onClick: async () => {
-              if (Auth.hasBillingScope()) {
-                setShowSpendReportConfigurationModal(true);
-              } else {
-                await authorizeAndLoadAccounts();
-                setShowSpendReportConfigurationModal(Auth.hasBillingScope());
-              }
-            },
-          },
-          [icon('edit', { size: 12 })]
-        ),
-        showSpendReportConfigurationModal &&
-          h(
-            Modal,
-            {
-              title: 'Configure Spend Reporting',
-              onDismiss: () => setShowSpendReportConfigurationModal(false),
-              okButton: h(
-                ButtonPrimary,
-                {
-                  disabled: !selectedDatasetProjectName || !selectedDatasetName,
-                  onClick: async () => {
-                    setShowSpendReportConfigurationModal(false);
-                    await updateSpendConfiguration(billingProject.projectName, selectedDatasetProjectName, selectedDatasetName);
-                  },
-                },
-                ['Ok']
-              ),
-            },
-            [
-              h(IdContainer, [
-                (id) =>
-                  h(Fragment, [
-                    h(FormLabel, { htmlFor: id, required: true }, ['Dataset Project ID']),
-                    h(TextInput, {
-                      id,
-                      onChange: setSelectedDatasetProjectName,
-                    }),
-                  ]),
-              ]),
-              h(IdContainer, [
-                (id) =>
-                  h(Fragment, [
-                    h(FormLabel, { htmlFor: id, required: true }, ['Dataset Name']),
-                    h(TextInput, {
-                      id,
-                      onChange: setSelectedDatasetName,
-                    }),
-                    div({ style: { marginTop: '1rem' } }, [
-                      ['See '],
-                      h(Link, { href: 'https://support.terra.bio/hc/en-us/articles/360037862771', ...Utils.newTabLinkProps }, ['our documentation']),
-                      [' for details on configuring spend reporting for billing projects.'],
-                    ]),
-                  ]),
-              ]),
-            ]
-          ),
-      ]),
-    !Auth.hasBillingScope() &&
-      div({ style: accountLinkStyle }, [
-        h(
-          Link,
-          {
-            onClick: authorizeAndLoadAccounts,
-          },
-          ['View billing account']
-        ),
-      ]),
-  ]);
-};
 
 const ProjectDetail = ({
   authorizeAndLoadAccounts,
@@ -373,7 +74,7 @@ const ProjectDetail = ({
   );
 
   const groups = groupByBillingAccountStatus(billingProject, workspacesInProject);
-  const billingAccountsOutOfDate = !(_.isEmpty(groups.error) && _.isEmpty(groups.updating));
+  const billingAccountsOutOfDate = true; //! (_.isEmpty(groups.error) && _.isEmpty(groups.updating));
 
   const tabToTable = {
     workspaces: h(Workspaces, {
@@ -461,7 +162,7 @@ const ProjectDetail = ({
         billingProject.projectName,
       ]),
       isGoogleBillingProject(billingProject) &&
-        h(GcpBillingAccountControls, {
+        h(BillingAccountControls, {
           authorizeAndLoadAccounts,
           billingAccounts,
           billingProject,
@@ -590,7 +291,7 @@ const ProjectDetail = ({
           removeUserFromBillingProject(deletingUser.roles, deletingUser.email).then(reloadBillingProject).then(reloadBillingProjectUsers);
         },
       }),
-    billingAccountsOutOfDate && h(BillingAccountSummaryPanel, { counts: _.mapValues(_.size, groups) }),
+    billingAccountsOutOfDate && h(BillingAccountSummary, { counts: _.mapValues(_.size, groups) }),
     updating && h(SpinnerOverlay, { mode: 'FullScreen' }),
   ]);
 };
